requirements.txt
Flask==1.1.1
PyQRCode==1.2.1
pypng==0.0.20
